+++ tests/unit/express/app.js
@@ -1,3 +1,5 @@
+var m                = require('module');
+var nconf            = require('nconf');
 var sinon            = require('sinon');
 var chai             = require('chai');
 var chaiAsPromised   = require('chai-as-promised');
@@ -7,9 +9,9 @@ var https            = require('https');
 var Express          = require('express');
 var logger           = require('bi-logger');
 var Session          = require('express-session');
-var Flash            = require('connect-flash');
 var CouchbaseODM     = require('kouchbase-odm');
 var ExpressValidator = require('bi-json-inspector');
+var BIServiceSDK     = require('bi-service-sdk').BIServiceSDK;
 
 var CouchbaseCluster = require('../../../lib/database/couchbase.js');
 var AppManager       = require('../../../lib/express/appManager.js');
@@ -242,7 +244,7 @@ describe('App', function() {
         });
 
         describe('useSession', function() {
-            it('should connect Session & Flash middlewares to express app', function() {
+            it('should connect Session middlewares to express app', function() {
                 var appUseSpy = sinon.spy(this.app, 'use');
                 var memcachedMock = new MemcachedStore();
 
@@ -252,7 +254,7 @@ describe('App', function() {
 
                 this.app.storage.session.should.be.equal(memcachedMock);
 
-                //TODO verify that actuall flash & session middleware was provided to the function
+                //TODO verify that actuall session middleware was provided to the function
                 sinon.assert.alwaysCalledWith(appUseSpy, sinon.match.func);
                 appUseSpy.calledTwice;
             });
@@ -338,6 +340,170 @@ describe('App', function() {
                 useSpy.should.have.been.calledWithExactly.apply(useSpy.should.have.been, args);
                 returnVal.should.be.equal(useSpy.getCall(0).returnValue);
             });
+        });
+
+        describe('useSDK', function() {
+            before(function() {
+                //
+                function SDKMock(options) {
+                    BIServiceSDK.call(this, options);
+
+                    this.version = 'v1.0';
+                }
+                SDKMock.prototype = Object.create(BIServiceSDK.prototype);
+                SDKMock.prototype.constructor = SDKMock;
+
+                //
+                function SDKMock2(options) {
+                    BIServiceSDK.call(this, options);
+
+                    this.version = 'v2.0';
+                }
+                SDKMock2.prototype = Object.create(BIServiceSDK.prototype);
+                SDKMock2.prototype.constructor = SDKMock2;
+
+                this.SDKMock  = SDKMock;
+                this.SDKMock2 = SDKMock2;
+
+                this.moduleRequireStub = sinon.stub(m.prototype, 'require');
+            });
+
+            after(function() {
+                this.moduleRequireStub.restore();
+            });
+
+            it('should connect provided SDK object to the app', function() {
+                var sdk = new this.SDKMock({baseURL: 'http://127.0.0.1'});
+                var key = 'sdk-name';
+
+                this.app.useSDK(key, sdk);
+
+                this.app.sdk.should.have.property(key);
+                this.app.sdk[key].should.have.property(sdk.version);
+                this.app.sdk[key][sdk.version].should.be.equal(sdk);
+            });
+
+            it('should connect all versions of a SDK to the app', function() {
+
+                var key = 'sdk-name';
+
+                this.configGetStub.withArgs(`services:${key}`).returns({
+                    npm: 'sdk-fake-pckg-name',
+                    host: '127.0.0.1',
+                    ssl: false
+                });
+
+                this.moduleRequireStub.withArgs('sdk-fake-pckg-name').returns({
+                    'v1.0': this.SDKMock,
+                    'v2.0': this.SDKMock2,
+                });
+
+                this.app.useSDK(key);
+
+                this.app.sdk.should.have.property(key);
+                this.app.sdk[key].should.have.property('v1.0');
+                this.app.sdk[key].should.have.property('v2.0');
+                this.app.sdk[key]['v1.0'].should.be.instanceof(this.SDKMock);
+                this.app.sdk[key]['v2.0'].should.be.instanceof(this.SDKMock2);
+            });
+
+            it('should return an object of connected sdk versions of related type', function() {
+                var key = 'sdk-name';
+
+                //connect one sdk version
+                var sdk = new this.SDKMock({baseURL: 'http://127.0.0.1'});
+
+                this.app.useSDK(key, sdk).should.be.eql({
+                    'v1.0': sdk
+                });
+
+                //connect another sdk version
+                this.configGetStub.withArgs(`services:${key}`).returns({
+                    npm: 'sdk-fake-pckg-name',
+                    host: '127.0.0.1',
+                    ssl: false
+                });
+
+                this.moduleRequireStub.withArgs('sdk-fake-pckg-name').returns({
+                    'v2.0': this.SDKMock2,
+                });
+
+                var versions = this.app.useSDK(key);
+                versions.should.have.property(sdk.version, sdk);
+                versions.should.have.property('v2.0').that.is.instanceof(this.SDKMock2);
+            });
+
+            it('should throw an Error when received sdk object is not instanceof BIServiceSDK', function() {
+                var self = this;
+
+                function testCase() {
+                    self.app.useSDK('sdk-name', Object.create({version: 'v1.0'}));
+                }
+
+                expect(testCase).to.throw(Error);
+            });
+
+            it('should throw an Error when required config option values are not found', function() {
+                var self = this;
+                var key = 'sdk-name';
+
+                //connect another sdk version
+                this.moduleRequireStub.withArgs('sdk-fake-pckg-name').returns({
+                    'v1.0': this.SDKMock,
+                });
+
+                this.configGetStub.withArgs(`services:${key}`).returns(null);
+
+                //npm services config object is required
+                expect(testCase).to.throw(Error);
+
+                this.configGetStub.withArgs(`services:${key}`).returns({
+                    npm: null,
+                    host: '127.0.0.1',
+                    ssl: false
+                });
+
+                //npm services config option is required
+                expect(testCase).to.throw(Error);
+
+                function testCase() {
+                    self.app.useSDK(key);
+                };
+            });
+
+            it('should not overwrite any already connected SDKs', function() {
+
+                var key = 'sdk-name';
+
+                this.configGetStub.withArgs(`services:${key}`).returns({
+                    npm: 'sdk-fake-pckg-name',
+                    host: '127.0.0.1',
+                    ssl: false
+                });
+
+                this.moduleRequireStub.withArgs('sdk-fake-pckg-name').returns({
+                    'v1.0': this.SDKMock,
+                    'v2.0': this.SDKMock2,
+                });
+
+                var sdk = new this.SDKMock({baseURL: 'http://127.0.0.1'});
+                var sdk2 = new this.SDKMock({baseURL: 'http://127.0.0.1'});
+
+
+                this.app.useSDK(key, sdk).should.be.eql({
+                    'v1.0': sdk
+                });
+
+                this.app.useSDK(key, sdk2).should.be.eql({
+                    'v1.0': sdk
+                });
+
+                var versions = this.app.useSDK(key);
+
+                versions.should.have.property('v1.0', sdk);
+                versions.should.have.property('v2.0').that.is.instanceof(this.SDKMock2);
+            });
+
         });
 
         describe('$buildExpressRouter', function() {
@@ -441,15 +607,26 @@ describe('App', function() {
                 }
             ].forEach(function(data, index) {
                 it(`should attach all routers to the root path when \`baseUrl\` config value is provided (${index})`, function() {
-                    this.configGetStub.withArgs('baseUrl').returns(data.baseUrl);
+                    var config = new nconf.Provider({
+                        store: {
+                            type: 'literal',
+                            store: {
+                                baseUrl: data.baseUrl
+                            }
+                        }
+                    });
 
-                    var expressUseSpy = sinon.spy(this.app.expressApp, 'use').withArgs(
+                    var app = this.appManager.buildApp(config, {
+                        name: Date.now() + index
+                    });
+
+                    var expressUseSpy = sinon.spy(app.expressApp, 'use').withArgs(
                         sinon.match.string,
                         sinon.match(this.matchers.expressRouter)
                     );
 
-                    this.app.buildRouter({url: data.routerUrl});
-                    this.app.build();
+                    app.buildRouter({url: data.routerUrl});
+                    app.build();
 
                     return this.nextTick(function() {
                         expressUseSpy.should.have.been.calledWith(data.expectedBinding);